requirements.txt
altgraph==0.17.4
bcrypt==4.3.0
cffi==1.17.1
colorama==0.4.6
cryptography==44.0.2
packaging==25.0
paramiko==3.5.1
pefile==2023.2.7
prompt_toolkit==3.0.51
pycparser==2.22
PyNaCl==1.5.0
pywin32-ctypes==0.2.3
PyYAML==6.0.2
setuptools==79.0.1
wcwidth==0.2.13